Exclude task with register from no-handler
Task that use when: previous.changes but register their result should
be excluded from the lint rule that forces a handler to be used.

An example of this scenario is:
- task 1: changes a file
- task 2: long running comamnd that is only need when task 1 performs a change
- task 3: does something else based on the output of 2

The reason not to use handlers here would be if task 3 is followed by
other tasks and needs to be run before the others.